+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Add `persist` field to the global config and individual recipes
   - Both default to `true`, but you can now set them to `false` to disable data persistence a single recipe, or all instances of the app. [See here for more](https://slumber.lucaspickering.me/book/user_guide/database.html#controlling-persistence)
 - Add actions to delete requests from the TUI
+  - Delete a single request from the history modal or the Request/Response pane
+  - Delete all requests for a recipe from the Recipe List/Recipe panes
 
 ### Changed